Add RiskOfRelics: FocusedConvergence compatibility
Risk of Relics: Focused Convergence relic causes boss chests to only present one
boss relic choice to the player, which is also repositioned to be in the middle.

If this is the case, we render the relic tooltip on the right-hand side like
normal, instead of on the left-hand side. We also position the Twitch vote text
underneath the relic properly now, which is something Risk of Relics didn't do.